requirements.txt
fastapi
pydantic
mysql-connector-python
uvicorn[standard]